problem-create-service/requirements.txt
spglib
booktype
aioredis
httpx
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0.0
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
alembic
asyncpg>=0.28.0
